isl/tests: bump coordinate types to 32bits

+++ src/intel/isl/tests/isl_tilememcpy_tiled_unittest.cpp
@@ -124,11 +124,11 @@ public:
    void TearDown();
    uint32_t swizzle_bitops(uint32_t num, uint8_t field,
                            uint8_t curr_ind, uint8_t swizzle_ind);
-   void bounded_byte_fill(uint8_t x1, uint8_t x2, uint8_t y1, uint8_t y2);
+   void bounded_byte_fill(uint32_t x1, uint32_t x2, uint32_t y1, uint32_t y2);
    void hex_oword_print(const uint8_t *buf, uint32_t size);
-   void convert_texture(uint8_t x1, uint8_t x2, uint8_t y1, uint8_t y2);
-   void compare_conv_result(uint8_t x1, uint8_t x2, uint8_t y1, uint8_t y2);
-   void run_test(uint8_t x1, uint8_t x2, uint8_t y1, uint8_t y2);
+   void convert_texture(uint32_t x1, uint32_t x2, uint32_t y1, uint32_t y2);
+   void compare_conv_result(uint32_t x1, uint32_t x2, uint32_t y1, uint32_t y2);
+   void run_test(uint32_t x1, uint32_t x2, uint32_t y1, uint32_t y2);
 };
 
 class tileYFixture : public tileTFixture,
@@ -186,7 +186,7 @@ void tileTFixture::TearDown()
    buf_dst = nullptr;
 }
 
-void tileTFixture::bounded_byte_fill(uint8_t x1, uint8_t x2, uint8_t y1, uint8_t y2)
+void tileTFixture::bounded_byte_fill(uint32_t x1, uint32_t x2, uint32_t y1, uint32_t y2)
 {
    uint8_t *itr = (uint8_t *) buf_src;
 
@@ -217,7 +217,7 @@ void tileTFixture::hex_oword_print(const uint8_t *buf, uint32_t size)
    }
 }
 
-void tileTFixture::convert_texture(uint8_t x1, uint8_t x2, uint8_t y1, uint8_t y2)
+void tileTFixture::convert_texture(uint32_t x1, uint32_t x2, uint32_t y1, uint32_t y2)
 {
    if (print_results) {
       printf("/************** Printing src ***************/\n");
@@ -243,8 +243,8 @@ void tileTFixture::convert_texture(uint8_t x1, uint8_t x2, uint8_t y1, uint8_t y
    }
 }
 
-void tileTFixture::compare_conv_result(uint8_t x1, uint8_t x2,
-                                       uint8_t y1, uint8_t y2)
+void tileTFixture::compare_conv_result(uint32_t x1, uint32_t x2,
+                                       uint32_t y1, uint32_t y2)
 {
    const struct isl_format_layout *fmtl = isl_format_get_layout(fmt);
    uint32_t x_max = (uint32_t) align(x2 * (fmtl->bpb / tile_info.format_bpb),
@@ -277,8 +277,8 @@ void tileTFixture::compare_conv_result(uint8_t x1, uint8_t x2,
    }
 }
 
-void tileTFixture::run_test(uint8_t x1, uint8_t x2,
-                            uint8_t y1, uint8_t y2)
+void tileTFixture::run_test(uint32_t x1, uint32_t x2,
+                            uint32_t y1, uint32_t y2)
 {
     bounded_byte_fill(x1, x2, y1, y2);
     convert_texture(x1, x2, y1, y2);
